refactor: inline pagination and move internal render helpers
