+++ src/components/DataTable/pagination/index.ts
@@ -1,2 +1,1 @@
-export * from './dataTablePagination';
 export * from './dataTablePagination.types';
