+++ src/components/DataTable/utils/renderCells.types.ts
